requirements.txt
streamlit
langchain.text_splitter 
langchain.embeddings.openai 
langchain.vectorstores
tqdm.auto
uuid